fix: improve perform by memo comp
