+++ apps/renderer/src/modules/claim/hooks.ts
@@ -6,20 +6,23 @@ import { getFeedById } from "~/store/feed"
 
 import { FeedClaimModalContent } from "./feed-claim-modal"
 
-export const useFeedClaimModal = ({ feedId }: { feedId?: string }) => {
+export const useFeedClaimModal = () => {
   const { present } = useModalStack()
   const { t } = useTranslation()
 
-  return useCallback(() => {
-    if (!feedId) return
+  return useCallback(
+    ({ feedId }: { feedId?: string }) => {
+      if (!feedId) return
 
-    const feed = getFeedById(feedId)
+      const feed = getFeedById(feedId)
 
-    if (!feed) return
+      if (!feed) return
 
-    present({
-      title: t("feed_claim_modal.title"),
-      content: () => createElement(FeedClaimModalContent, { feedId }),
-    })
-  }, [feedId, present])
+      present({
+        title: t("feed_claim_modal.title"),
+        content: () => createElement(FeedClaimModalContent, { feedId }),
+      })
+    },
+    [present, t],
+  )
 }
