+++ apps/renderer/src/queries/feed.ts
@@ -1,5 +1,8 @@
 import { formatXml } from "@follow/utils/utils"
 import { useMutation } from "@tanstack/react-query"
+import { useRef } from "react"
+import { useTranslation } from "react-i18next"
+import { toast } from "sonner"
 
 import { ROUTE_FEED_IN_FOLDER, ROUTE_FEED_PENDING } from "~/constants"
 import { useAuthQuery } from "~/hooks/common"
@@ -84,3 +87,28 @@ export const useRefreshFeedMutation = (feedId?: string) =>
       toastFetchError(err)
     },
   })
+
+export const useResetFeed = () => {
+  const { t } = useTranslation()
+  const toastIDRef = useRef<string | number | null>(null)
+
+  return useMutation({
+    mutationFn: async (feedId: string) => {
+      toastIDRef.current = toast.loading(t("sidebar.feed_actions.resetting_feed"))
+      await apiClient.feeds.reset.$get({ query: { id: feedId } })
+    },
+    onSuccess: (_, feedId) => {
+      entries.entries({ feedId }).invalidateRoot()
+      toast.success(
+        t("sidebar.feed_actions.reset_feed_success"),
+        toastIDRef.current ? { id: toastIDRef.current } : undefined,
+      )
+    },
+    onError: () => {
+      toast.error(
+        t("sidebar.feed_actions.reset_feed_error"),
+        toastIDRef.current ? { id: toastIDRef.current } : undefined,
+      )
+    },
+  })
+}
